Feat UI (#168)
* feat-ui

* fix round interval second

* remove HeroTable

+++ packages/contracts/mud.config.ts
@@ -244,8 +244,8 @@ export default mudConfig({
       schema: {
         seatNum: "uint8",
         withPassword: "bool",
-        createdAtBlock: "uint64",
-        updatedAtBlock: "uint64",
+        createdAtBlock: "uint32",
+        updatedAtBlock: "uint32",
         players: "address[]",
       },
     },
